fix: co-locate runtime readiness evidence
